backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
textstat==0.7.4
spacy==3.7.5
google-generativeai
python-dotenv==1.0.1
pytest==8.3.0
httpx==0.27.0
